24-Authorization-with-Policies

+++ app/Http/Controllers/ThreadsController.php
@@ -79,7 +79,9 @@ class ThreadsController extends Controller
     }
 
     public function destroy($channel, Thread $thread)
-    {
+    {   
+        $this->authorize('update', $thread);
+
         $thread->delete();
 
         if (request()->wantsJson()) {
